complete category CRUD

+++ src/categories/categories.controller.ts
@@ -6,40 +6,83 @@ import {
   Patch,
   Param,
   Delete,
+  Request,
+  UnauthorizedException,
+  ClassSerializerInterceptor,
+  UseGuards,
+  UseInterceptors,
+  BadRequestException,
 } from '@nestjs/common';
 import { CategoriesService } from './categories.service';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { UpdateCategoryDto } from './dto/update-category.dto';
+import { UsersService } from '../users/users.service';
+import { ApiTags, ApiBearerAuth } from '@nestjs/swagger';
+import { JwtAuthGuard } from '../auth/jwt-auth.guard';
 
+@ApiTags('Category')
+@ApiBearerAuth()
+@UseGuards(JwtAuthGuard)
+@UseInterceptors(ClassSerializerInterceptor)
 @Controller('categories')
 export class CategoriesController {
-  constructor(private readonly categoriesService: CategoriesService) {}
+  constructor(
+    private readonly categoriesService: CategoriesService,
+    private readonly usersService: UsersService,
+  ) {}
 
   @Post()
-  create(@Body() createCategoryDto: CreateCategoryDto) {
-    return this.categoriesService.create(createCategoryDto);
+  async create(@Body() createCategoryDto: CreateCategoryDto, @Request() req) {
+    const user = await this.usersService.findById(createCategoryDto.userId);
+
+    if (user.id !== req.user.id) {
+      throw new UnauthorizedException('Unable to create new category');
+    }
+    return await this.categoriesService.create(createCategoryDto, user);
   }
 
   @Get()
-  findAll() {
-    return this.categoriesService.findAll();
+  async findAll(@Request() req) {
+    return await this.categoriesService.findAll(req.user.id);
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.categoriesService.findOne(+id);
+  async findOne(@Param('id') id: string) {
+    return await this.categoriesService.findOne(+id);
   }
 
   @Patch(':id')
-  update(
-    @Param('id') id: string,
+  async update(
+    @Param('id') id: number,
     @Body() updateCategoryDto: UpdateCategoryDto,
+    @Request() req,
   ) {
-    return this.categoriesService.update(+id, updateCategoryDto);
+    // Check category is existing
+    const category = await this.categoriesService.findOne(id);
+
+    if (!category) {
+      throw new BadRequestException('Category does not exist.');
+    }
+
+    const user = await this.usersService.findById(req.user.id);
+
+    if (!user) {
+      throw new UnauthorizedException(
+        'You have no access to update this wallet',
+      );
+    }
+
+    return await this.categoriesService.update(+id, updateCategoryDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.categoriesService.remove(+id);
+  async remove(@Param('id') id: number) {
+    // Check category is existing
+    const category = await this.categoriesService.findOne(id);
+
+    if (!category) {
+      throw new BadRequestException('Category does not exist.');
+    }
+    return await this.categoriesService.remove(+id);
   }
 }
